% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_calib_and_d18O_boot.R
\name{apply_calib_and_d18O_boot}
\alias{apply_calib_and_d18O_boot}
\title{Calculate bootstrapped mean values for age, d18O, d13C, and D47 and calculate temperature and d18Osw}
\usage{
apply_calib_and_d18O_boot(data, calib, group, output = "summary", Nsim = NULL)
}
\description{
Calculate bootstrapped mean values for age, d18O, d13C, and D47 and calculate temperature and d18Osw
}
