% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_means.R
\name{bootstrap_means}
\alias{bootstrap_means}
\title{Calculate bootstrapped means for each group in a dataframe}
\usage{
bootstrap_means(data, group, age, d13C, d18O, D47, Nsim = 1e+05)
}
\arguments{
\item{data}{A dataframe or tibble that should have columns \code{outlier},
\code{broadid} ("other" for samples), \code{D47_final}, \code{d18O_PDB_vit},
\code{d13C_PDB_vit}, and the grouping column.}

\item{group}{The column name in \code{data} with a character or factor column
that contains the binning information that you want to calculate
bootstrapped averages for.}

\item{age}{The column name in \code{data} with the sample ages.}

\item{d13C}{The column name in \code{data} with the carbon isotope values.}

\item{d18O}{The column name in \code{data} with the oxygen isotope values.}

\item{D47}{The column name in \code{data} with the clumped isotope values.}

\item{Nsim}{The number of bootstraps you want to run. Defaults to the
number of rows in \code{calib}.}
}
\value{
A tibble with bootstrapped slope-intercept pair.
}
\description{
Calculate bootstrapped means for each group in a dataframe
}
