% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_temp_d18Osw.R
\name{calc_temp_d18Osw}
\alias{calc_temp_d18Osw}
\title{Calculate temperature and d18Osw}
\usage{
calc_temp_d18Osw(boot, calib, Nsim = NULL)
}
\arguments{
\item{calib}{A dataframe with draws from the bootstrapped (or Bayesian)
temperature regression. Should have columns \code{slope} and \code{intercept},
which are related via \code{clumpedr::revcal()}.}
}
\description{
From bootstrapped samples and a bootstrapped set of slope--intercept pairs.
}
