% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpedcalib-package.R
\docType{package}
\name{clumpedcalib-package}
\alias{clumpedcalib}
\alias{clumpedcalib-package}
\title{clumpedcalib: Calculate and Apply Clumped Isotope Calibrations Using Bootstrapping}
\description{
Calculate bootstrapped York regression from clumped isotope calibration data. Calculate bootstrapped sample means for the age, d13C, d18O, and D47. Apply the bootstrapped calibration to calculate temperature and d18Osw.
}
\author{
\strong{Maintainer}: Ilja Kocken \email{i.j.kocken@uu.nl} (\href{https://orcid.org/0000-0003-2196-8718}{ORCID})

}
\keyword{internal}
