% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18Osw_calc.R
\name{d18Osw_calc}
\alias{d18Osw_calc}
\title{Calculate the d18Osw from the d18Occ and temperature}
\usage{
d18Osw_calc(d18Occ, temperature, equation = NULL)
}
\arguments{
\item{d18Occ}{The oxygen isotope composition of the carbonate in VPDB.}

\item{temperature}{The formation temperature (in °C).}

\item{equation}{Character vector with the equation to use. Defaults to Kim & O'Neil, 1997.}
}
\value{
The oxygen isotope composition of the sea water in VSMOW.
}
\description{
Defaults to Kim & O'neil 1997 as modified by Bemis et al., 1998.
}
\references{
Kim, S.-T., & O’Neil, J. R. (1997). Equilibrium and nonequilibrium oxygen isotope effects in synthetic carbonates. Geochimica et Cosmochimica Acta, 61(16), 3461–3475. https://doi.org/10.1016/S0016-7037(97)00169-5

Bemis, B. E., Spero, H. J., Bijma, J., & Lea, D. W. (1998). Reevaluation of the oxygen isotopic composition of planktonic foraminifera: Experimental results and revised paleotemperature equations. Paleoceanography, 13(2), 150–160. https://doi.org/10.1029/98PA00070
}
\author{
Ilja J. Kocken
}
