% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_supported.R
\name{equation_supported}
\alias{equation_supported}
\title{Check if the provided equation is supported}
\usage{
equation_supported(equation)
}
\arguments{
\item{equation}{Character vector with equation reference. Must be one of \code{\link[=supported_equations]{supported_equations()}}.}
}
\value{
Character vector of cleaned up equation reference. Gives an error if it's currently not supported.
}
\description{
Also parses equation name into our internal simplified code.
See details for which equations are currently supported.
}
