% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d18Osw_calc.R
\name{general_bemis_d18Osw}
\alias{general_bemis_d18Osw}
\title{General Bemis et al., 1998 Table 1, inverted}
\usage{
general_bemis_d18Osw(d18Occ, temperature, a, b, c, d)
}
\arguments{
\item{d18Occ}{Input d18O of the calcite in VPDB.}

\item{temperature}{Input temperature.}

\item{a}{Intercept.}

\item{b}{Slope.}

\item{c}{Slope of quadratic term.}

\item{d}{The conversion correction from VSMOW to VPDB.}
}
\value{
The calculated \eqn{\delta^{18}O}{δ18O} value in VSMOW.
}
\description{
\eqn{\delta_{w} = \frac{\sqrt{-4ac + b^{2} + 4cT} + b + 2cT}{2c} - d}{δw = (sqrt(-4ac + b^2 + 4cT) + b +2cδc) / 2c - d}
}
\references{
Bemis, B. E., Spero, H. J., Bijma, J., & Lea, D. W. (1998). Reevaluation of the oxygen isotopic composition of planktonic foraminifera: Experimental results and revised paleotemperature equations. Paleoceanography, 13(2), 150–160. https://doi.org/10.1029/98PA00070
}
