% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calc.R
\name{general_bemis_temp}
\alias{general_bemis_temp}
\title{General Bemis et al. 1998 table 1}
\usage{
general_bemis_temp(d18Occ, d18Osw, a, b, c, d)
}
\arguments{
\item{d18Occ}{Input d18O of the calcite in VPDB.}

\item{d18Osw}{Input d18O of the sea water in VSMOW}

\item{a}{Intercept.}

\item{b}{Slope.}

\item{c}{Slope of quadratic term.}

\item{d}{The conversion correction from VSMOW to VPDB.}
}
\value{
The calculated temperature in °C.
}
\description{
\eqn{T (°C) = a + b(\delta_{c} - \delta_{w}) + c(\delta_{c}- \delta_{w})^{2}}{T (°C) = a + b(δc - δw) + c(δc - δw)^2}
}
\references{
Bemis, B. E., Spero, H. J., Bijma, J., & Lea, D. W. (1998). Reevaluation of the oxygen isotopic composition of planktonic foraminifera: Experimental results and revised paleotemperature equations. Paleoceanography, 13(2), 150–160. https://doi.org/10.1029/98PA00070
}
