% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_summary.R
\name{our_summary}
\alias{our_summary}
\title{Summarize the bootstrapped values into a mean, sd, and the 68\% and 95\% CIs}
\usage{
our_summary(boot, group)
}
\arguments{
\item{boot}{Output of \code{apply_calib_and_d18O_boot()}}

\item{group}{The group to summarize by.}
}
\description{
Summarize the bootstrapped values into a mean, sd, and the 68\% and 95\% CIs
}
