% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_d18Osw_calc.R
\name{temp_d18Osw_calc}
\alias{temp_d18Osw_calc}
\title{Calculate temperature and d18Osw}
\usage{
temp_d18Osw_calc(boot, calib, equation = NULL, Nsim = NULL)
}
\arguments{
\item{boot}{A tibble with bootstrapped slope-intercept pairs, output of \code{\link[=bootstrap_means]{bootstrap_means()}}.}

\item{calib}{A dataframe with draws from the bootstrapped (or Bayesian)
temperature regression. Should have columns \code{slope} and \code{intercept},
which are related via \code{clumpedr::revcal()}.}

\item{equation}{Character vector with the equation to use. Defaults to Kim & O'Neil, 1997.}
}
\value{
Same as boot but with additional columns \code{temp} and \code{d18Osw}.
}
\description{
From bootstrapped samples and a bootstrapped set of slope--intercept pairs.
}
