% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_supported.R
\name{equation_supported}
\alias{equation_supported}
\title{Parse equation name into our internal simplified code.}
\usage{
equation_supported(equation)
}
\description{
Throws an error if not supported.
}
