% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{filter_outliers}
\alias{filter_outliers}
\title{Filter clumped isotope outliers}
\usage{
filter_outliers(data, group)
}
\arguments{
\item{data}{A dataframe with columns \code{outlier}, \code{broadid},}
}
\description{
Filter clumped isotope outliers
}
