% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calc.R
\name{temp_calc}
\alias{temp_calc}
\title{Calculate the temperature from d18Occ and the d18Osw}
\usage{
temp_calc(d18Occ, d18Osw)
}
\arguments{
\item{d18Occ}{The d18O of the calcite in VPDB.}

\item{d18Osw}{The d18O of the sea water in VSMOW}
}
\value{
The temperature in degrees Celsius.
}
\description{
This is the relationship from Kim & O'neil 1997,
as updated by Bemis et al., 1998
}
