% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calc.R
\name{temp_calc}
\alias{temp_calc}
\title{Calculate the temperature from d18Occ and the d18Osw}
\usage{
temp_calc(d18Occ, d18Osw, equation = NULL)
}
\arguments{
\item{d18Occ}{The oxygen isotope composition of the carbonate in VPDB.}

\item{d18Osw}{The oxygen isotope composition of the sea water in VSMOW}

\item{equation}{Character vector with the equation to use. Defaults to Kim & O'Neil, 1997.}
}
\value{
The temperature in degrees Celsius.
}
\description{
This is the relationship from Kim & O'neil 1997,
as updated by Bemis et al., 1998
}
\author{
Ilja J. Kocken
}
